requirements.txt
pytorch
numpy
pandas